backsolteo/requirements.txt
alembic==1.13.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
docopt==0.6.2
Flask==3.0.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.1
MarkupSafe==2.1.4
pipreqs==0.4.13
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.41.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
WTForms==3.1.2
WTForms-JSON==0.3.5
yarg==0.1.9
